requirements.txt
anyio==3.6.2
black==22.10.0
click==8.1.3
fastapi==0.87.0
h11==0.14.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
sniffio==1.3.0
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
